move dialog row component out
code duplication
