+++ lib/v2/screens/transfer/send/send_confirmation/components/custom_transaction_success_dialog.dart
@@ -5,7 +5,6 @@ import 'package:intl/intl.dart';
 import 'package:seeds/v2/datasource/remote/model/custom_transaction_model.dart';
 import 'package:seeds/v2/design/app_theme.dart';
 import 'package:seeds/v2/components/custom_dialog.dart';
-import 'package:seeds/v2/components/profile_avatar.dart';
 import 'package:seeds/v2/constants/app_colors.dart';
 import 'package:seeds/v2/i18n/transfer/transfer.i18n.dart';
 
@@ -122,45 +121,6 @@ class CustomTransactionSuccessDialog extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class DialogRow extends StatelessWidget {
-  final String? imageUrl;
-  final String account;
-  final String? name;
-  final String? toOrFromText;
-
-  const DialogRow({Key? key, this.imageUrl, required this.account, this.name, this.toOrFromText}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        ProfileAvatar(size: 60, image: imageUrl, account: account, nickname: name),
-        Expanded(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 16, right: 8),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(name ?? account, textAlign: TextAlign.start, style: Theme.of(context).textTheme.button),
-                const SizedBox(height: 8),
-                Text(account, style: Theme.of(context).textTheme.subtitle2LowEmphasis)
-              ],
-            ),
-          ),
-        ),
-        Container(
-          decoration: const BoxDecoration(
-              borderRadius: BorderRadius.all(Radius.elliptical(4, 4)), color: AppColors.lightGreen6),
-          child: Padding(
-            padding: const EdgeInsets.only(top: 4, bottom: 4, right: 8, left: 8),
-            child: Text(toOrFromText!, style: Theme.of(context).textTheme.subtitle2),
-          ),
-        ),
-      ],
     );
   }
 }
